Remove --tx-mempool-mode and the TxMemPool implementation
Rip out the nonce-aware transaction mempool introduced in #971 now that
--tx-batch-mode is the only supported path going forward. Retains the
NonceProvider/NonceView interfaces, LocalNonceProvider, and shared error
types that BatchTxPool depends on.

- Delete services/requester/tx_mempool.go and its unit tests
- Delete tests/tx_mempool_test.go
- Move the fastPathSubmitTimeout and flushReasonFastPath/Prefix
  constants from tx_mempool.go into tx_pool.go, since BatchTxPool
  imports them; drop the TxMemPool-only flushReasonTTL
- Remove config fields TxMemPoolMode, TxCollectionWindow,
  TxSubmissionSpacing, TxPoolTTL, TxMaxBatchSize, TxMaxNonceGap
- Remove the six --tx-mempool-* / --tx-max-* CLI flags and the
  TxMemPoolMode validation block from cmd/run/cmd.go
- Remove the TxMemPoolMode branch from bootstrap/bootstrap.go
- Remove ErrNonceTooLow and ErrNonceTooHigh (only referenced by the
  removed TxMemPool)
- Move fundEOA into tests/tx_batching_test.go and the fakeNonceProvider
  + signedTestTx helpers into services/requester/batch_tx_pool_test.go
  so the remaining tests keep compiling
- Drop the mempool rows from README's Configuration Flags table
