+++ config/config.go
@@ -129,37 +129,6 @@ type Config struct {
 	// of the events from the sealed block in the Flow network.
 	// CAUTION: This feature is experimental and will cause the node to halt if the events don't match.
 	ExperimentalSealingVerificationEnabled bool
-	// TxMemPoolMode configures the gateway to use the transaction mempool:
-	// transactions carrying the expected next nonce (with nothing in flight)
-	// are submitted immediately, out-of-order transactions are held until their
-	// nonce gap fills, and consecutive Cadence submissions for the same EOA are
-	// spaced apart to avoid Collection Node re-ordering.
-	TxMemPoolMode bool
-	// TxCollectionWindow is the per-EOA sliding collection window used by the
-	// transaction mempool. The window resets on each new transaction arrival
-	// from the same EOA; when it elapses, the collected transactions are flushed.
-	TxCollectionWindow time.Duration
-	// TxSubmissionSpacing is the minimum gap between two consecutive Cadence
-	// transaction submissions for the same EOA (recommended ~1.5x the block
-	// production rate). It also serves as the flush deadline for a
-	// continuously-fed collection window, anchored at first enqueue.
-	TxSubmissionSpacing time.Duration
-	// TxPoolTTL is how long the transaction mempool holds an out-of-order
-	// transaction waiting for its nonce gap to fill. On expiry the transaction
-	// is submitted anyway, so the failure is observable instead of a silent drop.
-	TxPoolTTL time.Duration
-	// TxMaxBatchSize is the maximum number of EVM transactions submitted in a
-	// single EVM.batchRun Cadence transaction by the transaction mempool,
-	// bounded by the Cadence transaction computation limit.
-	TxMaxBatchSize int
-	// TxMaxNonceGap is how far ahead of an EOA's on-chain nonce the transaction
-	// mempool will accept a nonce. A transaction whose nonce exceeds
-	// indexedNonce + TxMaxNonceGap is rejected up front with ErrNonceTooHigh,
-	// giving the client immediate feedback instead of holding an unexecutable tx
-	// until TTL. 0 means no upper bound (any future nonce is accepted). This
-	// bounds only the upper end: a nonce below the indexed nonce is always
-	// rejected with ErrNonceTooLow regardless of this setting.
-	TxMaxNonceGap uint64
 	// RpcRequestTimeout is the maximum duration at which JSON-RPC requests should generate
 	// a response, before they timeout.
 	RpcRequestTimeout time.Duration
